Reimplementation of CollapsibleHeaderLayout using stickyHeader.
Measuring expansion progress of the layout is tricky as we need to know its
minimum and maximum possible heights, which we can't do without actually rendering it.

Here we use the initial measured (maximum) height and subsequent scroll events
to calculate an approximate minimum height via a system of linear equations.

When the user reaches the actual minimum height we can use that measured height as the
canonical minimum height value for later calculations of expansion progress.

This works much better than previously attempted approaches such as forcing the view to
minimum height (causing a brief flicker), or hiding the view entirely while the limits
are measured.